fix(autoDiscover): skip files index.html already references as module scripts
createDirectiveClientAutoDiscover added every src/* file with a top-of-
file "use client" as an explicit Vite input. When that overlaps with the
module-script src in index.html (the conventional case: src/client.tsx
referenced by <script type="module" src="/src/client.tsx">), Vite
dedupes its own index.html manifest entry away. processCssFilesForPages
then loses the entry's transitive css and the rendered HTML drops the
imported global styles.

Read the project's index.html once at autoDiscover time, skip any
file the scan would have added that index.html's <script type="module">
already references. Vite continues to discover those modules via the
index.html input on its own, and the manifest's index.html entry
survives with its css intact.

Reverts the disk-read fallback briefly added to processCssFilesForPages
on this branch.

Verified against mmc with no explicit clientEntry: manifest now has
the index.html entry with its globalStyles css, /8mmc/index.html links
the stylesheet, 357/357 unit tests pass.
