+++ plugin/react-static/processCssFilesForPages.ts
@@ -6,7 +6,6 @@ import type {
 import type { Logger, Manifest } from "vite";
 import { collectManifestCss } from "../helpers/collectManifestCss.js";
 import { createUnifiedCssProcessor } from "../helpers/createUnifiedCssProcessor.js";
-import { resolveIndexHtmlClientEntry } from "../helpers/resolveIndexHtmlClientEntry.js";
 import { join } from "node:path";
 
 interface ProcessCssFilesForPagesOptions {
@@ -31,23 +30,11 @@ export function processCssFilesForPages({
 } {
   const cssFilesByPage = new Map();
 
-  // Collect the global styles reachable from the project's client entry.
-  // Try the manifest's `"index.html"` key first; if absent, walk every
-  // `<script type="module" src>` in the on-disk index.html.
-  let indexHtmlCssInputs = collectManifestCss(
+  // First collect global styles from index.html
+  const indexHtmlCssInputs = collectManifestCss(
     staticManifest ?? {},
     "index.html"
   );
-  if (
-    Object.keys(indexHtmlCssInputs).length === 0 &&
-    userOptions.projectRoot
-  ) {
-    const fallbackEntries = resolveIndexHtmlClientEntry(userOptions.projectRoot);
-    for (const entry of fallbackEntries) {
-      const fromEntry = collectManifestCss(staticManifest ?? {}, entry);
-      indexHtmlCssInputs = { ...indexHtmlCssInputs, ...fromEntry };
-    }
-  }
   const clientEntryCssInputs = userOptions.clientEntry
     ? collectManifestCss(
         staticManifest ?? {},
